requirements.txt
Django==3.2.10
matplotlib==3.4.3
nltk==3.6.3
numpy==1.22.4
PyPDF2==1.26.0
python-docx==0.8.11